backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
alembic==1.13.2
python-dotenv==1.0.1
email-validator==2.2.0

